入院患者数: Rename components/DashedRectangleTimeBarChart.vue to components/index/CardsMonitoring/HospitalizedNumber/Chart.vue
