+++ components/index/CardsMonitoring/HospitalizedNumber/Card.vue
@@ -1,7 +1,7 @@
 <template>
   <v-col cols="12" md="6" class="DataCard HospitalizedNumberCard">
     <client-only>
-      <dashed-rectangle-time-bar-chart
+      <chart
         :title="$t('モニタリング項目(6)')"
         :title-id="'number-of-hospitalized'"
         :info-titles="[$t('入院患者数')]"
@@ -33,19 +33,19 @@
             </li>
           </ul>
         </template>
-      </dashed-rectangle-time-bar-chart>
+      </chart>
     </client-only>
   </v-col>
 </template>
 
 <script>
-import DashedRectangleTimeBarChart from '@/components/DashedRectangleTimeBarChart.vue'
+import Chart from '@/components/index/CardsMonitoring/HospitalizedNumber/Chart.vue'
 import positiveStatus from '@/data/positive_status.json'
 import formatGraph from '@/utils/formatGraph'
 
 export default {
   components: {
-    DashedRectangleTimeBarChart,
+    Chart,
   },
   data() {
     const formatData = positiveStatus.data
