+++ components/index/CardsMonitoring/HospitalizedNumber/Chart.vue
